store_dashboard/requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.2.10
djangorestframework==3.14.0
dnspython==2.6.1
et-xmlfile==1.1.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.7
magic-filter==1.0.12
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
openpyxl==3.1.3
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.45
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pymongo==4.7.2
pyOpenSSL==24.1.0
pytest==8.2.1
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zope.interface==6.4.post2
